+++ src/parser/syntaxes/common.ts
@@ -6,6 +6,7 @@ import { parseExpr } from './expressions.js';
 
 import type { ITokenStream } from '../streams/token-stream.js';
 import type * as Ast from '../../node.js';
+import { parseType } from './types.js';
 
 /**
  * ```abnf
@@ -134,16 +135,6 @@ export function parseBlock(s: ITokenStream): (Ast.Statement | Ast.Expression)[] 
 	return steps;
 }
 
-//#region Type
-
-export function parseType(s: ITokenStream): Ast.TypeSource {
-	if (s.is(TokenKind.At)) {
-		return parseFnType(s);
-	} else {
-		return parseNamedType(s);
-	}
-}
-
 /**
  * ```abnf
  * OptionalSeparator = [SEP]
@@ -167,79 +158,3 @@ export function parseOptionalSeparator(s: ITokenStream): boolean {
 		}
 	}
 }
-
-/**
- * ```abnf
- * FnType = "@" "(" ParamTypes ")" "=>" Type
- * ParamTypes = [Type *(SEP Type)]
- * ```
-*/
-function parseFnType(s: ITokenStream): Ast.TypeSource {
-	const startPos = s.getPos();
-
-	s.expect(TokenKind.At);
-	s.next();
-	s.expect(TokenKind.OpenParen);
-	s.next();
-
-	const params: Ast.TypeSource[] = [];
-	while (!s.is(TokenKind.CloseParen)) {
-		if (params.length > 0) {
-			switch (s.getTokenKind()) {
-				case TokenKind.Comma: {
-					s.next();
-					break;
-				}
-				case TokenKind.EOF: {
-					throw new AiScriptUnexpectedEOFError(s.getPos());
-				}
-				default: {
-					throw new AiScriptSyntaxError('separator expected', s.getPos());
-				}
-			}
-		}
-		const type = parseType(s);
-		params.push(type);
-	}
-
-	s.expect(TokenKind.CloseParen);
-	s.next();
-	s.expect(TokenKind.Arrow);
-	s.next();
-
-	const resultType = parseType(s);
-
-	return NODE('fnTypeSource', { params, result: resultType }, startPos, s.getPos());
-}
-
-/**
- * ```abnf
- * NamedType = IDENT ["<" Type ">"]
- * ```
-*/
-function parseNamedType(s: ITokenStream): Ast.TypeSource {
-	const startPos = s.getPos();
-
-	let name: string;
-	if (s.is(TokenKind.Identifier)) {
-		name = s.getTokenValue();
-		s.next();
-	} else {
-		s.expect(TokenKind.NullKeyword);
-		s.next();
-		name = "null";
-	}
-
-	// inner type
-	let inner: Ast.TypeSource | undefined;
-	if (s.is(TokenKind.Lt)) {
-		s.next();
-		inner = parseType(s);
-		s.expect(TokenKind.Gt);
-		s.next();
-	}
-
-	return NODE('namedTypeSource', { name, inner }, startPos, s.getPos());
-}
-
-//#endregion Type
